+++ R/ard_survival_survfit.R
@@ -324,7 +324,7 @@ extract_multi_strata <- function(x, df_stat) {
       warning = list(NULL),
       error = list(NULL)
     ) %>%
-    structure(., class = c("card", class(.))) %>%
+    cards::as_card() %>%
     cards::tidy_ard_column_order() %>%
     cards::tidy_ard_row_order()
 }