+++ tests/testthat/test-ard_stats_wilcox_test_onesample.R
@@ -47,6 +47,6 @@ test_that("ard_stats_wilcox_test_onesample() works", {
       cards::ADSL,
       variables = character(0)
     ),
-    dplyr::tibble()
+    dplyr::tibble() |> cards::as_card()
   )
 })